BETA 1.6 PRERELEASE - Theoretical Optimizations
- Massively changed enemy update code, causing heavy optimizations.
- Lost an rebuild research menu, so it's differnet from previous
version.
- Added module mod interface, though basic with unfinished GUI, this
allows the player to better control the behaviour of individual modules.
- Disabled assembler builder scene temporarely, it'll return.
- Fixed a whole lotta bugs, and broke even more stuff.
- Changed the internal workings of the Research data structure.
- Added research tree backup, to avoid loosing it again.
- Other stuff that I've forgotten about.

From the previous versions since 1.5
- Added assembly selection screen. This will limit the player in 10
assemblies of his own choosing to use on the battlefield, though the
number is softcoded and can be easily changed.
- Worked on getting custom battlefields working.
- Massively improved multiple filesaving structures by using internal
.NET serialization instead of custom files.
- Added a square, apparently.

The current module mods are limited to bases, and include 2 mods:
- Sorting - Defines which order enemies are chosen as targets, defualt
is "First", and includes "Closest", "Furthest", "Strogest", "Weakest",
"First", "Last", "Random".

- Priorty - Defines if the base should either prioritize or completely
ignore the different colors of enemy.

I'm terribly sorry for whatever spelling mistakes I make right now, it
is a stressful time right now.