BETA 1.6.4 - Last One for Now
- Fixed research menu buttons images getting more and more invisible
each time you purchased research.
- Fixed locked or researched research buttons being clickable.
- Some missles now turn sharper if they're pointing away from their
target.
- Fixed Research Collector not being purchaseable. Other special modules
are yet to return (Slowfield generator)
- Added Turbo Mode, which when activated will double the games speed -
THIS FEATURE IS EXPERIMENTAL.
- Cleaned up purchase menu code.
- Added cool ass animations to various buttons in the game, as an
experiement towards BETA 1.7

This is likely the last patch to come before BETA 1.7, unless something
urgent needs to be fixed. BETA 1.7 is planned to include a bunch of
asthetic improvements overall, and generally to polish up the game.
There are still a few glaring bugs remaining, including the Research
Collector not really being compatable with a lot of the code. Though,
that will have to wait.

- Lomztein