BUILD 8 - Assembly Editor is back, baby!
- Readded assembly editor scene, still needs work though.
- Added another new main menu scene, much happier with this background.
- Added animation to darkOverlay fade in and out.
- Fixed assembly saving failing to save names and descriptions.
- Changed some if(Game.game) to if(Game.currentScene).

The entire thing with supporting all these customizeable features is
quite tiring to implement. It could be argued that it adds a bit of
gameplay for like, 5% of players, though I doubt I'll be implementing it
in many other future games. It's the kind of thing that requires a lot
of planning, and having never done customization of this scale, I lacked
that. Perhaps the experience gained from this project will lead to
better customization in the future.