1.8 PRE-GOLD - I lied.
In the end I decided not to rework the entire goddamn game. Have you seen this code? It's closer to spaghetti than my dinner last night. Which, to be fair, was rice.

 * Changed how research points drop from enemies. Now drops much more evenly throughout the wave.
 * Fixed flamethrower being useless.
 * Nerfed wave 12.
 * Research Collector now has practically infinite range.
 * Rewrote some descriptions, fixing typos and cringy writing.
 * Removed all but one graphics setting, since they're useless anyways.
 * Fixed issues with DPS calculator.
 * Fixed healthbars being visible for a single frame when spawning.
 * Fixed highscores not being sorted correctly.
 * Fixed sniper rifle weapon having the wrong projectile.
 * Ported to the latest version of Unity, 2017.2.
 * Fixed walls and build flashes being completely black.
 * Fixed assembly selection menu sorting.
 * Removed arrow from wall texture.
 * Fixed some formatting of assembly context menu.
 * Fixed assembly context menu not updating with kills.
 * Fixed larger class modules being weird with placement.
 * Long Range Base is now class 4
 * Fixed slowfield generator purchase button sprite being out of wack.
 * Minor hud improvements.

Theres not gonna come any new features from now on, and the next patch will be the final one, containing balance changes and bug fixes, because balance is kind of out of wack right now.

After that, the game is done. Gold version 1.0.